+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * (improvement) Add `StoryblokBackendUrlGenerator::generateStoryEditUrlById()`.
 * (improvement) Simplify diff generation.
 * (improvement) Don't sync structure to Storyblok if not in prod.
+* (improvement) Add new RichTextStyling option `RichTextStyling::UnsetFormatting`.
+* (improvement) Add support for transforming HardBreak, CodeBlock, HorizontalRule and Emoji in `RichTextTransformer`.
 
 
 3.10.0